+++ changelog.md
@@ -8,6 +8,8 @@ Codion Change Log
 - Database.USER_PROPERTY and PASSWORD_PROPERTY renamed USER and PASSWORD.
 ### is.codion.common.model
 - SummaryModel.SummaryValues.Factory.createSummaryValues() renamed create().
+### is.codion.common.reactive
+- State.Builder.group() now supports multiple groups.
 ### is.codion.swing.common.ui
 - MenuItemBuilder, unused type parameter removed.
 - Control.copy() return type fixed.